Fix the OSDK CI failure
