+++ kernel/aster-nix/src/lib.rs
@@ -41,9 +41,6 @@ extern crate alloc;
 extern crate lru;
 #[macro_use]
 extern crate controlled;
-#[cfg(ktest)]
-#[macro_use]
-extern crate ktest;
 #[macro_use]
 extern crate getset;
