+++ kernel/aster-nix/src/vm/vmar/options.rs
@@ -136,7 +136,7 @@ impl<R> VmarChildOptions<R> {
 #[cfg(ktest)]
 mod test {
     use aster_rights::Full;
-    use ostd::mm::VmIo;
+    use ostd::{mm::VmIo, prelude::*};
 
     use super::*;
     use crate::vm::{
