+++ kernel/comps/network/Cargo.toml
@@ -13,7 +13,6 @@ bitflags = "1.3"
 bitvec = { version = "1.0.1", default-features = false, features = ["alloc"]}
 component = { path = "../../libs/comp-sys/component" }
 int-to-c-enum = { path = "../../libs/int-to-c-enum" }
-ktest = { path = "../../../ostd/libs/ktest" }
 log = "0.4"
 ostd = { path = "../../../ostd" }
 pod = { git = "https://github.com/asterinas/pod", rev = "d7dba56" }
