+++ kernel/libs/aster-util/Cargo.toml
@@ -12,6 +12,5 @@ typeflags-util = { path = "../typeflags-util" }
 aster-rights-proc = { path = "../aster-rights-proc" }
 aster-rights = { path = "../aster-rights" }
 inherit-methods-macro = { git = "https://github.com/asterinas/inherit-methods-macro", rev = "98f7e3e" }
-ktest = { path = "../../../ostd/libs/ktest" }
 
 [features]
