+++ kernel/libs/aster-util/src/coeff.rs
@@ -127,9 +127,10 @@ impl Mul<u32> for Coeff {
 
 #[cfg(ktest)]
 mod test {
-    use ktest::ktest;
+    use ostd::prelude::*;
 
     use super::*;
+
     #[ktest]
     fn calculation() {
         let coeff = Coeff::new(23456, 56789, 1_000_000_000);
