--- ostd/libs/ktest-proc-macro/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "ktest-proc-macro"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1.0.69"
-quote = "1.0.33"
-rand = "0.8.5"
-syn = { version = "2.0.29", features = ["full"] }
