+++ ostd/libs/ktest/Cargo.toml
@@ -7,4 +7,3 @@ edition = "2021"
 
 [dependencies]
 owo-colors = "3.5.0"
-ktest-proc-macro = { path = "../ktest-proc-macro" }
