+++ ostd/libs/ostd-macros/Cargo.toml
@@ -11,4 +11,5 @@ proc-macro = true
 [dependencies]
 proc-macro2 = "1.0.78"
 quote = "1.0.35"
+rand = "0.8.5"
 syn = { version = "2.0.48", features = ["full"] }
